Rewrite dockerfile to use multi-stage build
By using a multi-stage build we reduce the image size from 957 MB to
500 MB. We also reduce the size of the layers that need to be pushed
after making a minimal change to the source code to just the size of
the jar file (about 30 MB). Previously hundreds of MB of layers had
to be pushed every time (at least partly because we downloaded Gradle
every time).

We also add two environment variables that can be set to pass extra
options to the JVM and to Gerrit Herald.

Finally, by using exec in the image entrypoint we make sure the JVM
ends up as pid 1 so it gets the SIGTERM signal when the container
is stopped.