Make path to property file configurable (#10)
Welding the configuration into the Docker image doesn't fit everyone,
and up until now it hasn't been possible to change the configuration in
config.properties in any other way. Introduce a herald.properties system
property that can be set to the file system path of the property file to
load. If the property is unset it'll fall back to the config.properties
file in the jar file, i.e. this change is backwards compatible.