Add property to control whether exchange declarations are active (#8)
Previously all exchange declarations were made with
Channel.exchangeDeclare(), i.e. an active declaration that
requires that the user has the configure permission for the
exchange in question. Because that permission allows the
user to delete the exchange it's rarely granted except to
broker administrators.

The new S_EXCHANGE_ACTIVE boolean property controls whether
exchange declarations are active or passive.